LayoutTests/imported/w3c/web-platform-tests/css/css-writing-modes/text-orientation-parsing-sideways-right-001-expected.txt


FAIL -webkit-text-orientation: sideways-right assert_equals: expected (string) "sideways" but got (undefined) undefined
